Add menu item, add auth file

+++ VB/UDisk/UDisk/Form1.vb
@@ -17,7 +17,7 @@ Public Class Form1
     Public Const DBT_USERDEFINED = &HFFFF
 
     Protected Overrides Sub WndProc(ByRef m As System.Windows.Forms.Message)
-        If m.Msg = WM_DEVICECHANGE Then
+        If m.Msg = WM_DEVICECHANGE And Menu_Enable.CheckState = CheckState.Checked Then
             Select Case m.WParam
                 Case WM_DEVICECHANGE
                 Case DBT_DEVICEARRIVAL 'U盘插入
@@ -32,17 +32,31 @@ Public Class Form1
 
                                 AutorunFiles = RemoveableDevices.ToString + "Autorun.inf"
 
-                                If My.Computer.FileSystem.FileExists(AutorunFiles) Then
-                                    My.Computer.FileSystem.DeleteFile(AutorunFiles, FileIO.UIOption.OnlyErrorDialogs, FileIO.RecycleOption.DeletePermanently)
-                                    NotifyIconAutoRun.ShowBalloonTip(3000, "发现威胁", "文件 " + AutorunFiles + " 已删除", ToolTipIcon.Warning)
+                                ''''''''''''''''''''''''
+                                'Using fs As New System.IO.FileStream(AutorunFiles, IO.FileMode.Open)
+                                'Using md5 As System.Security.Cryptography.MD5 = System.Security.Cryptography.MD5.Create()
+                                'Dim hash As Byte() = md5.ComputeHash(fs)
+                                'NotifyIconAutoRun.ShowBalloonTip(300, "Hash", BitConverter.ToString(hash), ToolTipIcon.Info)
+                                'MsgBox(BitConverter.ToString(hash))
+                                'End Using
+                                'End Using
+                                '''''''''''''''''''''''''''
+                                Dim FileReader As System.IO.StreamReader
 
+                                FileReader = My.Computer.FileSystem.OpenTextFileReader(AutorunFiles)
+                                Dim FirstString As String = FileReader.ReadLine()
+                                If My.Computer.FileSystem.FileExists(AutorunFiles) And FirstString <> ";DOSSTONED Authorised" Then
+                                    FileReader.Close()
+                                    My.Computer.FileSystem.DeleteFile(AutorunFiles, FileIO.UIOption.OnlyErrorDialogs, FileIO.RecycleOption.DeletePermanently)
+                                    NotifyIconAutoRun.ShowBalloonTip(300, "发现威胁", "文件 " + AutorunFiles + " 已删除", ToolTipIcon.Warning)
+                                ElseIf FirstString = ";DOSSTONED Authorised" Then
+                                    NotifyIconAutoRun.ShowBalloonTip(300, "U盘安全", RemoveableDevices.ToString.Remove(1, 2) + "盘已经通过DOSSTONED认证，可放心使用", ToolTipIcon.Info)
                                 End If
+                                FileReader.Close()
 
                                 If My.Computer.FileSystem.DirectoryExists(AutorunFiles) Then
                                     My.Computer.FileSystem.DeleteDirectory(AutorunFiles, FileIO.UIOption.OnlyErrorDialogs, FileIO.RecycleOption.DeletePermanently)
-
-                                    'MsgBox("目录 " + AutorunFiles + " 已删除")
-                                    NotifyIconAutoRun.ShowBalloonTip(3000, "发现异常", "目录 " + AutorunFiles + " 已删除", ToolTipIcon.Warning)
+                                    NotifyIconAutoRun.ShowBalloonTip(300, "发现异常", "目录 " + AutorunFiles + " 已删除", ToolTipIcon.Warning)
                                 End If
 
                             End If
@@ -69,7 +83,26 @@ Public Class Form1
         MyBase.WndProc(m)
     End Sub
 
+    Private Sub ToolStripMenu_Exit_Click(ByVal sender As System.Object, ByVal e As System.EventArgs)
+        Application.Exit()
+    End Sub
+
+    Private Sub Menu_Exit_Click(ByVal sender As System.Object, ByVal e As System.EventArgs) Handles Menu_Exit.Click
+        Application.Exit()
+    End Sub
+
+
+    Private Sub Menu_Enable_Click(ByVal sender As System.Object, ByVal e As System.EventArgs) Handles Menu_Enable.Click
+
+        If Menu_Enable.CheckState = CheckState.Checked Then
+            NotifyIconAutoRun.ShowBalloonTip(300, "AutoRun 已保护", "您的U盘处于保护之中", ToolTipIcon.Info)
+            NotifyIconAutoRun.Text = "AutoRun 已保护"
+        Else
+            NotifyIconAutoRun.ShowBalloonTip(300, "AutoRun 未保护", "您的U盘未处于保护之中", ToolTipIcon.Warning)
+            NotifyIconAutoRun.Text = "AutoRun 未保护"
+        End If
+
+    End Sub
 End Class
 
 
-'本文来自CSDN博客，转载请标明出处：http://blog.csdn.net/wzuomin/archive/2007/07/27/1711720.aspx
